feat: upgrade controller.js to controller.ts without errors
